Confirmação de remoção

+++ curso_dev_kotlin/Convidados/app/src/main/java/com/example/convidados/view/viewholder/GuestsViewHolder.kt
@@ -1,6 +1,8 @@
 package com.example.convidados.view.viewholder
 
+import android.content.DialogInterface
 import android.view.View
+import androidx.appcompat.app.AlertDialog
 import androidx.recyclerview.widget.RecyclerView
 import com.example.convidados.databinding.RowGuestBinding
 import com.example.convidados.model.GuestModel
@@ -17,7 +19,17 @@ class GuestsViewHolder(private val bind: RowGuestBinding, private val listener: 
         }
 
         bind.textName.setOnLongClickListener {
-            listener.onDelete(guest.id)
+
+            AlertDialog.Builder(itemView.context)
+                .setTitle("Remoção de convidado")
+                .setMessage("Tem certeza que deseja remover?")
+                .setPositiveButton("Sim") { dialog, which ->
+                    listener.onDelete(guest.id)
+                }
+                .setNegativeButton("Não", null )
+                .create()
+                .show()
+
             true
         }
     }
